Create GameMap and GameEngine to encapsulate some of towerdefense.cs 's responsibilities
