+++ TowerDefense/StaticClasses/GameStats.cs
@@ -3,11 +3,17 @@ namespace TowerDefense
 {
     static class GameStats
     {
-        public static int Gold { get; set; }
+        internal static int Gold { get; set; }
+        internal static bool AttackPhase { get; set;}
+        internal static bool PlayerLoses { get; set; }
+        internal static int Level { get; set; }
 
         static GameStats()
         {
             Gold = Constants.StartingGold;
+            AttackPhase = false;
+            PlayerLoses = false;
+            Level = 1;
         }
     }
 }
