+++ TowerDefense/TowerDefense.cs
@@ -14,18 +14,12 @@ namespace TowerDefense
         SpriteBatch batch;
         MouseHandler mouse;
         Viewport viewport;
+
+        GameEngine gameEngine = new GameEngine();
+        GameMap gameMap = new GameMap();
+        List<Button> buttonlist = new List<Button>();
         Button startButton;
         Button upgradeButton;
-        List<Button> buttonlist = new List<Button>();
-        List<Tower> towerlist = new List<Tower>();
-        List<Enemy> enemylist = new List<Enemy>();
-        List<Projectile> projectilelist = new List<Projectile>();
-        List<FloatingText> floatingTextList = new List<FloatingText>();
-        Node[,] nodes = new Node[Constants.MapSize.X + 1, Constants.MapSize.Y + 1];
-        bool attackPhase = false;
-        bool playerLoses = false;
-        int level = 0;
-        double lastSpawnedTime = 0;
 
         public TowerDefense()
         {
@@ -59,88 +53,15 @@ namespace TowerDefense
             buttonlist.Add(new Button(new Point(10 + 64, (int)(viewport.Height * .5f)), ResourceManager.Portal, HoveringContext.ButtonPortal));
             buttonlist.Add(new Button(new Point(10 + 16, (int)(viewport.Height * .56f)), ResourceManager.Cheese, HoveringContext.ButtonCheese));
 
-            CreateMap();
-
             mouse = new MouseHandler(Point.Zero, ResourceManager.DefaultCursor);
         }
-
-        //This map stuff should eventually live elsewhere
-        private void CreateMap()
-        {
-            int actualY = Constants.MapStart.Y;
-            int actualX = Constants.MapStart.X;
-            for (int i = 0; i <= Constants.MapSize.X; i++)
-            {
-                for (int j = 0; j <= Constants.MapSize.Y; j++)
-                {
-                    nodes[i, j] = new Node(new Point(actualX, actualY), new Point(i, j), ResourceManager.Grass);
-                    actualY += Constants.NodeSize.Y;
-                }
-                actualY = Constants.MapStart.Y;
-                actualX += Constants.NodeSize.X;
-            }
-        }
-
-        private bool MouseInGameBounds()
-        {
-            int topY = Constants.MapStart.Y;
-            int leftX = Constants.MapStart.X;
-
-            return mouse.pos.Y > topY && mouse.pos.X > leftX &&
-                        mouse.pos.Y < topY + (Constants.MapSize.Y * Constants.NodeSize.Y) &&
-                        mouse.pos.X < leftX + (Constants.MapSize.X * Constants.NodeSize.X);
-        }
-        //
 
         protected override void Update(GameTime gameTime)
         {
             HandleMouse();
             mouse.Update();
-            if (attackPhase)
-            {
-                foreach (Enemy e in enemylist)
-                {
-                    if (!e.spawned && (gameTime.TotalGameTime.TotalSeconds - lastSpawnedTime) > e.spawnRate)
-                    {
-                        e.spawn();
-                        lastSpawnedTime = gameTime.TotalGameTime.TotalSeconds;
-                    }
-                }
-                List<Enemy> temp = new List<Enemy>();
-                foreach (Enemy e in enemylist)
-                {
-                    if (e.dead)
-                    {
-                        temp.Add(e);
-                    }
-                    e.move();
-                    if (e.lose)
-                    {
-                        MessageLog.GameOver();
-                        playerLoses = true;
-                    }
-                }
-                foreach (Enemy e in temp)
-                {
-                    enemylist.Remove(e);
-                }
-                foreach (Tower t in towerlist)
-                {
-                    projectilelist = t.Attack(enemylist, projectilelist, gameTime.TotalGameTime.TotalSeconds, (d, p) =>
-                    {
-                        floatingTextList.Add(new FloatingText(p, ResourceManager.GameFont, d.ToString()));
-                    });
-                }
-                projectilelist.RemoveAll(p => p.Move());
-            }
-            if (enemylist.Count == 0 && attackPhase)
-            {
-                attackPhase = false;
-                projectilelist.Clear();
-                MessageLog.LevelComplete(level * 2 + (int)(GameStats.Gold * .05f), level);
-                GameStats.Gold = GameStats.Gold + (level * 2 + (int)(GameStats.Gold * .05f));
-            }
-            floatingTextList.RemoveAll(f => f.Update(gameTime));
+            gameEngine.Update(gameTime);
+
             base.Update(gameTime);
         }
 
@@ -149,20 +70,11 @@ namespace TowerDefense
             GraphicsDevice.Clear(Color.WhiteSmoke);
 
             batch.Begin();
-            for (int i = 0; i <= Constants.MapSize.X; i++)
-            {
-                for (int j = 0; j <= Constants.MapSize.Y; j++)
-                {
-                    nodes[i, j].Draw(batch);
-                }
-            }
 
+            gameMap.Draw(batch);
             startButton.Draw(batch);
-            enemylist.ForEach(e => e.Draw(batch));
             buttonlist.ForEach(b => b.Draw(batch));
-            towerlist.ForEach(t => t.Draw(batch));
-            projectilelist.ForEach(p => p.Draw(batch));
-            floatingTextList.ForEach(f => f.Draw(batch));
+            gameEngine.Draw(batch);
 
             if (mouse.SelectionContext == SelectionContext.TowerSelected)
             {
@@ -191,7 +103,7 @@ namespace TowerDefense
 
         private void HandleMouse()
         {
-            if (playerLoses)
+            if (GameStats.PlayerLoses)
             {
                 if (mouse.MouseState.LeftButton == ButtonState.Pressed)
                 {
@@ -200,12 +112,13 @@ namespace TowerDefense
                 return;
             }
 
-            if (mouse.MouseState.LeftButton == ButtonState.Pressed && startButton.ContainsPoint(mouse.pos) && !attackPhase && mouse.SelectionContext == SelectionContext.None)
+            if (mouse.MouseState.LeftButton == ButtonState.Pressed && startButton.ContainsPoint(mouse.pos) && !GameStats.AttackPhase && mouse.SelectionContext == SelectionContext.None)
             {
-                StartLevel();
+                mouse.UpdateTex(ResourceManager.DefaultCursor);
+                gameEngine.StartLevel();
             }
             HandleMouseHover();
-            if (!attackPhase)
+            if (!GameStats.AttackPhase)
             {
                 if (mouse.MouseState.LeftButton == ButtonState.Pressed)
                 {
@@ -259,28 +172,11 @@ namespace TowerDefense
                 mouse.HoveringContext = upgradeButton.HoveringContext;
             }
 
-            enemylist.ForEach(e =>
-            {
-                e.Hovering = e.BoundingBox().Contains(mouse.pos);
-                if (e.Hovering)
-                {
-                    mouse.HoveredObject = e;
-                    mouse.HoveringContext = HoveringContext.Enemy;
-                }
-            });
+            gameEngine.HandleMouseHover(mouse);
+            gameMap.HandleMouseHover(mouse);
 
-            if (!attackPhase)
+            if (!GameStats.AttackPhase)
             {
-                foreach (Node n in nodes)
-                {
-                    n.Hovering = n.BoundingBox().Contains(mouse.pos) && mouse.SelectionContext != SelectionContext.PlacingTower;
-                    if (n.Hovering)
-                    {
-                        mouse.HoveredObject = n;
-                        mouse.HoveringContext = n.wall || n.portal || n.cheese ? HoveringContext.FilledNode : HoveringContext.EmptyNode;
-                    }
-                }
-
                 buttonlist.ForEach(b =>
                 {
                     b.Hovering = b.BoundingBox().Contains(mouse.pos);
@@ -290,46 +186,6 @@ namespace TowerDefense
                         mouse.HoveringContext = b.HoveringContext;
                     }
                 });
-                towerlist.ForEach(t =>
-                {
-                    t.Hovering = t.BoundingBox().Contains(mouse.pos) && mouse.SelectionContext == SelectionContext.None;
-                    if (t.Hovering)
-                    {
-                        mouse.HoveredObject = t;
-                        mouse.HoveringContext = HoveringContext.Tower;
-                    }
-                });
-            }
-        }
-
-        private void StartLevel()
-        {
-            mouse.UpdateTex(ResourceManager.DefaultCursor);
-            attackPhase = true;
-            level++;
-            MessageLog.Level(level);
-            Random rand = new Random();
-            double num = rand.NextDouble();
-            if (num < .3)
-            {
-                for (int i = 0; i < (15 + level); i++)
-                {
-                    enemylist.Add(new Enemy(level * 8, 2, ResourceManager.Enemy, nodes, "Malaria", 1.0f, .4f));
-                }
-            }
-            else if (num < .6)
-            {
-                for (int i = 0; i < (30 + 2 * level); i++)
-                {
-                    enemylist.Add(new Enemy(level * 4, 2, ResourceManager.Enemy, nodes, "Tuberculosis", .33f, .2f));
-                }
-            }
-            else
-            {
-                for (int i = 0; i < (5 + level / 2); i++)
-                {
-                    enemylist.Add(new Enemy(level * 16, 1, ResourceManager.Enemy, nodes, "AIDS", 1.5f, .8f));
-                }
             }
         }
 
@@ -339,25 +195,9 @@ namespace TowerDefense
             {
                 mouse.HoveredObject.HandleLeftClick(mouse);
             }
-            if (mouse.SelectionContext == SelectionContext.PlacingTower && MouseInGameBounds())
-            {
-                Tower t = mouse.SelectedObject as Tower;
-                if (GameStats.Gold >= t.Cost)
-                {
-                    GameStats.Gold = GameStats.Gold - t.Cost;
-                    towerlist.Add(t);
-                    t.Position = mouse.pos;
-                    mouse.SelectedObject = null;
-                    mouse.SelectionContext = SelectionContext.None;
-                    mouse.UpdateTex(ResourceManager.DefaultCursor);
-                    ResourceManager.WallSound.Play();
-                }
-                else
-                {
-                    MessageLog.NotEnoughGold();
-                }
-            }
-            else if (mouse.SelectionContext == SelectionContext.PlacingPortalEntrance && mouse.HoveringContext == HoveringContext.EmptyNode)
+            gameEngine.HandleLeftClick(mouse);
+
+            if (mouse.SelectionContext == SelectionContext.PlacingPortalEntrance && mouse.HoveringContext == HoveringContext.EmptyNode)
             {
                 Node n = mouse.HoveredObject as Node;
                 n.portal = true;
@@ -410,23 +250,13 @@ namespace TowerDefense
 
         private void HandleRightClick()
         {
+            gameEngine.HandleRightClick(mouse);
+
             if (mouse.SelectionContext == SelectionContext.PlacingPortalExit)
             {
                 mouse.PortalEntrance.portal = false;
                 mouse.PortalEntrance.defaultSet();
                 mouse.PortalEntrance = null;
-            }
-            else if (mouse.SelectionContext == SelectionContext.TowerSelected)
-            {
-                Tower t = mouse.SelectedObject as Tower;
-                t.Selected = false;
-            }
-            else if (mouse.HoveringContext == HoveringContext.Tower && mouse.SelectionContext == SelectionContext.None)
-            {
-                Tower t = mouse.HoveredObject as Tower;
-                GameStats.Gold = GameStats.Gold + t.Cost;
-                towerlist.Remove(t);
-                ResourceManager.SellSound.Play();
             }
             else if (mouse.HoveringContext == HoveringContext.FilledNode && mouse.SelectionContext == SelectionContext.None)
             {
